+++ tests/testthat/test-roxytest-testexamples-02-pipeline.R
@@ -2,7 +2,7 @@
 
 # File R/"02-pipeline.R": @testexamples
 
-test_that("[unknown alias] @ L114", {
+test_that("[unknown alias] @ L115", {
   
   library(bregr)
   # 1. Pipeline -------------------------